backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
alembic 